requirements.txt
prometheus_client>=0.6.0
semver >= 2.7.0
pyyaml == 6.0.2
